use command-exists instead of manually checking if cmd is avalible
this commit replaces old code that was designed to determine if things like rustup or wasm-gc were installed
the new implementation uses the npm package command-exists, which gives us better async, errs, etc

+++ packages/react-scripts/scripts/utils/rustUtils.js
@@ -1,5 +1,6 @@
 const execSync = require('child_process').execSync;
-const chalk = require('chalk');
+const chalk = require('react-dev-utils/chalk');
+const commandExistsSync = require('command-exists').sync;
 
 module.exports = {
   build: () => {
@@ -13,19 +14,14 @@ module.exports = {
       console.log(chalk.bold.red('error compiling rust'));
     }
   },
-  isRustInstalled: () => {
-    try {
-      execSync('rustup show');
-      return true;
-    } catch (error) {
-      return false;
-    }
-  },
+  isRustInstalled: () => commandExistsSync('rustup'),
   installRustWebAssemblyTools: () => {
-    try {
-      execSync('cargo install wasm-gc', { stdio: 'inherit' });
-    } catch {
-      console.log(`${chalk.bold.red('error installing wasm-gc')} please install manually ${chalk.red('cargo install wasm-gc')}')`);
+    if (!commandExistsSync('wasm-gc')) {
+      try {
+        execSync('cargo install wasm-gc', { stdio: 'inherit' });
+      } catch {
+        console.log(`${chalk.bold.red('error installing wasm-gc')} please install manually ${chalk.red('cargo install wasm-gc')}')`);
+      }
     }
     execSync(
       'rustup target add wasm32-unknown-unknown --toolchain nightly',
